refactor: add radix-ui/tabs library for tabs
will break the way tabs component are used
