+++ package.json
@@ -85,6 +85,7 @@
         "@radix-ui/react-dialog": "^0.1.5",
         "@radix-ui/react-radio-group": "^1.0.0",
         "@radix-ui/react-switch": "^0.1.5",
+        "@radix-ui/react-tabs": "^1.0.0",
         "@stitches/react": "^1.2.6",
         "@storybook/addon-a11y": "^6.5.9",
         "classnames": "^2.3.1"
