+++ src/components/core/tabs/stories/tabs.stories.tsx
@@ -1,7 +1,5 @@
-import type { Story } from '@storybook/react';
-import Tab from '@core/tabs/tab';
-import type { TabsProps } from '@core/tabs/tabs';
-import Tabs from '@core/tabs/tabs';
+import { Story } from '@storybook/react';
+import Tabs, { TabsProps } from '../tabs';
 import DepositIconLight from './assets/ic-deposit-light.svg';
 import WithdrawIconLight from './assets/ic-withdraw-light.svg';
 import TransferIconLight from './assets/ic-transfer-light.svg';
@@ -14,9 +12,6 @@ import TransactionIconDark from './assets/ic-transactions-dark.svg';
 export default {
     title: 'Tabs',
     argTypes: {
-        active_index: {
-            description: 'Sets the active tab index.',
-        },
         contained: {
             description: 'If set to `true`, the tab background will be changed.',
             defaultValue: false,
@@ -25,7 +20,6 @@ export default {
                 defaultValue: { summary: false },
             },
         },
-
         size: {
             control: {
                 type: 'select',
@@ -46,32 +40,37 @@ export default {
 
 const Template: Story<TabsProps> = (args, { globals: { theme } }) => {
     const isDark = theme === 'dark';
-    return (
-        <Tabs {...args}>
-            <Tab icon={isDark ? DepositIconDark : DepositIconLight} label="Deposit">
-                Deposit
-            </Tab>
-            <Tab icon={isDark ? WithdrawIconDark : WithdrawIconLight} label="Withdraw">
-                Withdraw
-            </Tab>
-            <Tab icon={isDark ? TransferIconDark : TransferIconLight} label="Transfer">
-                Transfer
-            </Tab>
-            <Tab icon={isDark ? TransactionIconDark : TransactionIconLight} label="Transactions">
-                Transactions
-            </Tab>
-        </Tabs>
-    );
+    const tabs = [
+        {
+            label: 'Deposit',
+            icon: isDark ? DepositIconDark : DepositIconLight,
+            content: 'Deposit',
+        },
+        {
+            label: 'Withdraw',
+            icon: isDark ? WithdrawIconDark : WithdrawIconLight,
+            content: 'Withdraw',
+        },
+        {
+            label: 'Transfer',
+            icon: isDark ? TransferIconDark : TransferIconLight,
+            content: 'Transfer',
+        },
+        {
+            label: 'Transactions',
+            icon: isDark ? TransactionIconDark : TransactionIconLight,
+            content: 'Transactions',
+        },
+    ];
+    return <Tabs {...args} tabs={tabs} default_selected="Withdraw" />;
 };
 
 export const Bordered = Template.bind({});
 Bordered.args = {
-    active_index: 0,
     contained: false,
 };
 
 export const Contained = Template.bind({});
 Contained.args = {
-    active_index: 0,
     contained: true,
 };
